requirements.txt
aniso8601==9.0.1
blinker==1.5
click==8.1.3
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
